fixed int cast

+++ Entities/PlayerEntity.cs
@@ -98,7 +98,7 @@ namespace SprintZero1.Entities
         public void TakeDamage(int damage)
         {
            
-                HUDManager.decrementHearts(damage,_playerHealth);
+                HUDManager.decrementHearts(damage,(int)_playerHealth);
             
             _playerHealth -= damage;
